Refactoring: Review Comments
